Add episode handling
For each episode create an episode menu which features some technical
data and, more importantly, the summary of the episode, as supplied in
the `.txt` files for each one.

Optimised the code for many of the repetitive tasks by using an Object
containing the three states and the corresponding colours.

Also reduced the `ffmpeg` output by adding some flags to hide the banner
and reduce the log level. Finally, added a small reminder of what to do
next.